pyproject.toml
[tool.poetry]
name = "fasttutorial"
version = "0.1.0"
description = ""
authors = ["Babar Rasheed <babarcs511@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
databases = "^0.9.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
